Making facet menus hierarchical again
